requirements.txt
Django
djangorestframework
djongo
gunicorn
